Backport sidecar affinity support
