+++ caas/kubernetes/provider/application/application.go
@@ -17,7 +17,6 @@ import (
 	"github.com/juju/errors"
 	"github.com/juju/featureflag"
 	"github.com/juju/loggo"
-	"github.com/juju/utils/v3/arch"
 	"github.com/juju/version/v2"
 	"github.com/kr/pretty"
 	appsv1 "k8s.io/api/apps/v1"
@@ -1240,24 +1239,6 @@ func (a *app) applicationPodSpec(config caas.ApplicationConfig) (*corev1.PodSpec
 		return containers[i].Name < containers[j].Name
 	})
 
-	resourceRequests := corev1.ResourceList(nil)
-	millicores := 0
-	if config.Constraints.HasCpuPower() {
-		if resourceRequests == nil {
-			resourceRequests = corev1.ResourceList{}
-		}
-		// 100 cpu power is 100 millicores.
-		millicores = int(*config.Constraints.CpuPower)
-		resourceRequests[corev1.ResourceCPU] = *resource.NewMilliQuantity(int64(millicores), resource.DecimalSI)
-	}
-	if config.Constraints.HasMem() {
-		if resourceRequests == nil {
-			resourceRequests = corev1.ResourceList{}
-		}
-		bytes := *config.Constraints.Mem * 1024 * 1024
-		resourceRequests[corev1.ResourceMemory] = *resource.NewQuantity(int64(bytes), resource.BinarySI)
-	}
-
 	env := []corev1.EnvVar{
 		{
 			Name:  "JUJU_CONTAINER_NAMES",
@@ -1339,10 +1320,6 @@ func (a *app) applicationPodSpec(config caas.ApplicationConfig) (*corev1.PodSpec
 				SubPath:   "charm/containers",
 			},
 		},
-		Resources: corev1.ResourceRequirements{
-			Requests: resourceRequests,
-			Limits:   resourceRequests,
-		},
 	}}
 
 	imagePullSecrets := []corev1.LocalObjectReference(nil)
@@ -1416,9 +1393,6 @@ func (a *app) applicationPodSpec(config caas.ApplicationConfig) (*corev1.PodSpec
 					SubPath:   fmt.Sprintf("charm/containers/%s", v.Name),
 				},
 			},
-			Resources: corev1.ResourceRequirements{
-				Limits: resourceRequests,
-			},
 		}
 		if v.Image.Password != "" {
 			imagePullSecrets = append(imagePullSecrets, corev1.LocalObjectReference{Name: a.imagePullSecretName(v.Name)})
@@ -1426,31 +1400,10 @@ func (a *app) applicationPodSpec(config caas.ApplicationConfig) (*corev1.PodSpec
 		containerSpecs = append(containerSpecs, container)
 	}
 
-	nodeSelector := map[string]string(nil)
-	if config.Constraints.HasArch() {
-		cpuArch := *config.Constraints.Arch
-		cpuArch = arch.NormaliseArch(cpuArch)
-		// Convert to Golang arch string
-		switch cpuArch {
-		case arch.AMD64:
-			cpuArch = "amd64"
-		case arch.ARM64:
-			cpuArch = "arm64"
-		case arch.PPC64EL:
-			cpuArch = "ppc64le"
-		case arch.S390X:
-			cpuArch = "s390x"
-		default:
-			return nil, errors.NotSupportedf("architecture %q", cpuArch)
-		}
-		nodeSelector = map[string]string{"kubernetes.io/arch": cpuArch}
-	}
-
 	automountToken := true
-	return &corev1.PodSpec{
+	spec := &corev1.PodSpec{
 		AutomountServiceAccountToken:  &automountToken,
 		ServiceAccountName:            a.serviceAccountName(),
-		NodeSelector:                  nodeSelector,
 		ImagePullSecrets:              imagePullSecrets,
 		TerminationGracePeriodSeconds: pointer.Int64Ptr(300),
 		InitContainers: []corev1.Container{{
@@ -1529,7 +1482,12 @@ func (a *app) applicationPodSpec(config caas.ApplicationConfig) (*corev1.PodSpec
 				},
 			},
 		},
-	}, nil
+	}
+	err := ApplyConstraints(spec, a.name, config.Constraints, ConfigureConstraint)
+	if err != nil {
+		return nil, errors.Annotate(err, "processing constraints")
+	}
+	return spec, nil
 }
 
 func (a *app) ensureImagePullSecrets(applier resources.Applier, config caas.ApplicationConfig) error {
