+++ caas/kubernetes/provider/k8s.go
@@ -8,7 +8,6 @@ import (
 	"context"
 	"fmt"
 	"regexp"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
@@ -41,6 +40,7 @@ import (
 	"k8s.io/utils/pointer"
 
 	"github.com/juju/juju/caas"
+	k8sapplication "github.com/juju/juju/caas/kubernetes/provider/application"
 	"github.com/juju/juju/caas/kubernetes/provider/constants"
 	"github.com/juju/juju/caas/kubernetes/provider/resources"
 	k8sspecs "github.com/juju/juju/caas/kubernetes/provider/specs"
@@ -52,7 +52,6 @@ import (
 	k8sannotations "github.com/juju/juju/core/annotations"
 	"github.com/juju/juju/core/assumes"
 	coreconfig "github.com/juju/juju/core/config"
-	"github.com/juju/juju/core/constraints"
 	"github.com/juju/juju/core/devices"
 	"github.com/juju/juju/core/paths"
 	coreresources "github.com/juju/juju/core/resources"
@@ -811,244 +810,6 @@ func (k *kubernetesClient) DeleteService(appName string) (err error) {
 	return nil
 }
 
-func processConstraints(pod *core.PodSpec, appName string, cons constraints.Value) error {
-	// TODO(caas): Allow constraints to be set at the container level.
-	if mem := cons.Mem; mem != nil {
-		if err := configureConstraint(pod, "memory", fmt.Sprintf("%dMi", *mem)); err != nil {
-			return errors.Annotatef(err, "configuring memory constraint for %s", appName)
-		}
-	}
-	if cpu := cons.CpuPower; cpu != nil {
-		if err := configureConstraint(pod, "cpu", fmt.Sprintf("%dm", *cpu)); err != nil {
-			return errors.Annotatef(err, "configuring cpu constraint for %s", appName)
-		}
-	}
-	nodeSelector := map[string]string(nil)
-	if cons.HasArch() {
-		cpuArch := *cons.Arch
-		cpuArch = arch.NormaliseArch(cpuArch)
-		// Convert to Golang arch string
-		switch cpuArch {
-		case arch.AMD64:
-			cpuArch = "amd64"
-		case arch.ARM64:
-			cpuArch = "arm64"
-		case arch.PPC64EL:
-			cpuArch = "ppc64le"
-		case arch.S390X:
-			cpuArch = "s390x"
-		default:
-			return errors.NotSupportedf("architecture %q", cpuArch)
-		}
-		nodeSelector = map[string]string{"kubernetes.io/arch": cpuArch}
-	}
-	if pod.NodeSelector != nil {
-		for k, v := range nodeSelector {
-			pod.NodeSelector[k] = v
-		}
-	} else if nodeSelector != nil {
-		pod.NodeSelector = nodeSelector
-	}
-
-	// Translate tags to pod or node affinity.
-	// Tag names are prefixed with "pod.", "anti-pod.", or "node."
-	// with the default being "node".
-	// The tag 'topology-key', if set, is used for the affinity topology key value.
-	if cons.Tags != nil {
-		affinityLabels := make(map[string]string)
-		for _, labelPair := range *cons.Tags {
-			parts := strings.Split(labelPair, "=")
-			if len(parts) != 2 {
-				return errors.Errorf("invalid affinity constraints: %v", affinityLabels)
-			}
-			key := strings.Trim(parts[0], " ")
-			value := strings.Trim(parts[1], " ")
-			affinityLabels[key] = value
-		}
-
-		if err := processNodeAffinity(pod, affinityLabels); err != nil {
-			return errors.Annotatef(err, "configuring node affinity for %s", appName)
-		}
-		if err := processPodAffinity(pod, affinityLabels); err != nil {
-			return errors.Annotatef(err, "configuring pod affinity for %s", appName)
-		}
-	}
-	if cons.Zones != nil {
-		zones := *cons.Zones
-		affinity := pod.Affinity
-		if affinity == nil {
-			affinity = &core.Affinity{
-				NodeAffinity: &core.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &core.NodeSelector{
-						NodeSelectorTerms: []core.NodeSelectorTerm{{}},
-					},
-				},
-			}
-			pod.Affinity = affinity
-		}
-		nodeSelector := &affinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0]
-		nodeSelector.MatchExpressions = append(nodeSelector.MatchExpressions,
-			core.NodeSelectorRequirement{
-				Key:      "failure-domain.beta.kubernetes.io/zone",
-				Operator: core.NodeSelectorOpIn,
-				Values:   zones,
-			})
-	}
-	return nil
-}
-
-const (
-	podPrefix      = "pod."
-	antiPodPrefix  = "anti-pod."
-	topologyKeyTag = "topology-key"
-	nodePrefix     = "node."
-)
-
-func processNodeAffinity(pod *core.PodSpec, affinityLabels map[string]string) error {
-	affinityTags := make(map[string]string)
-	for key, value := range affinityLabels {
-		keyVal := key
-		if strings.HasPrefix(key, "^") {
-			if len(key) == 1 {
-				return errors.Errorf("invalid affinity constraints: %v", affinityLabels)
-			}
-			key = key[1:]
-		}
-		if strings.HasPrefix(key, podPrefix) || strings.HasPrefix(key, antiPodPrefix) {
-			continue
-		}
-		key = strings.TrimPrefix(keyVal, nodePrefix)
-		affinityTags[key] = value
-	}
-
-	updateSelectorTerms := func(nodeSelectorTerm *core.NodeSelectorTerm, tags map[string]string) {
-		// Sort for stable ordering.
-		var keys []string
-		for k := range tags {
-			keys = append(keys, k)
-		}
-		sort.Strings(keys)
-		for _, tag := range keys {
-			allValues := strings.Split(tags[tag], "|")
-			for i, v := range allValues {
-				allValues[i] = strings.Trim(v, " ")
-			}
-			op := core.NodeSelectorOpIn
-			if strings.HasPrefix(tag, "^") {
-				tag = tag[1:]
-				op = core.NodeSelectorOpNotIn
-			}
-			nodeSelectorTerm.MatchExpressions = append(nodeSelectorTerm.MatchExpressions, core.NodeSelectorRequirement{
-				Key:      tag,
-				Operator: op,
-				Values:   allValues,
-			})
-		}
-	}
-	var nodeSelectorTerm core.NodeSelectorTerm
-	updateSelectorTerms(&nodeSelectorTerm, affinityTags)
-	if pod.Affinity == nil {
-		pod.Affinity = &core.Affinity{}
-	}
-	pod.Affinity.NodeAffinity = &core.NodeAffinity{
-		RequiredDuringSchedulingIgnoredDuringExecution: &core.NodeSelector{
-			NodeSelectorTerms: []core.NodeSelectorTerm{nodeSelectorTerm},
-		},
-	}
-	return nil
-}
-
-func processPodAffinity(pod *core.PodSpec, affinityLabels map[string]string) error {
-	affinityTags := make(map[string]string)
-	antiAffinityTags := make(map[string]string)
-	for key, value := range affinityLabels {
-		notVal := false
-		if strings.HasPrefix(key, "^") {
-			if len(key) == 1 {
-				return errors.Errorf("invalid affinity constraints: %v", affinityLabels)
-			}
-			notVal = true
-			key = key[1:]
-		}
-		if !strings.HasPrefix(key, podPrefix) && !strings.HasPrefix(key, antiPodPrefix) {
-			continue
-		}
-		if strings.HasPrefix(key, podPrefix) {
-			key = strings.TrimPrefix(key, podPrefix)
-			if notVal {
-				key = "^" + key
-			}
-			affinityTags[key] = value
-		}
-		if strings.HasPrefix(key, antiPodPrefix) {
-			key = strings.TrimPrefix(key, antiPodPrefix)
-			if notVal {
-				key = "^" + key
-			}
-			antiAffinityTags[key] = value
-		}
-	}
-	if len(affinityTags) == 0 && len(antiAffinityTags) == 0 {
-		return nil
-	}
-
-	updateAffinityTerm := func(affinityTerm *core.PodAffinityTerm, tags map[string]string) {
-		// Sort for stable ordering.
-		var keys []string
-		for k := range tags {
-			keys = append(keys, k)
-		}
-		sort.Strings(keys)
-		var (
-			labelSelector v1.LabelSelector
-			topologyKey   string
-		)
-		for _, tag := range keys {
-			if tag == topologyKeyTag {
-				topologyKey = tags[tag]
-				continue
-			}
-			allValues := strings.Split(tags[tag], "|")
-			for i, v := range allValues {
-				allValues[i] = strings.Trim(v, " ")
-			}
-			op := v1.LabelSelectorOpIn
-			if strings.HasPrefix(tag, "^") {
-				tag = tag[1:]
-				op = v1.LabelSelectorOpNotIn
-			}
-			labelSelector.MatchExpressions = append(labelSelector.MatchExpressions, v1.LabelSelectorRequirement{
-				Key:      tag,
-				Operator: op,
-				Values:   allValues,
-			})
-		}
-		affinityTerm.LabelSelector = &labelSelector
-		if topologyKey != "" {
-			affinityTerm.TopologyKey = topologyKey
-		}
-	}
-	if pod.Affinity == nil {
-		pod.Affinity = &core.Affinity{}
-	}
-	var affinityTerm core.PodAffinityTerm
-	updateAffinityTerm(&affinityTerm, affinityTags)
-	if len(affinityTerm.LabelSelector.MatchExpressions) > 0 {
-		pod.Affinity.PodAffinity = &core.PodAffinity{
-			RequiredDuringSchedulingIgnoredDuringExecution: []core.PodAffinityTerm{affinityTerm},
-		}
-	}
-
-	var antiAffinityTerm core.PodAffinityTerm
-	updateAffinityTerm(&antiAffinityTerm, antiAffinityTags)
-	if len(antiAffinityTerm.LabelSelector.MatchExpressions) > 0 {
-		pod.Affinity.PodAntiAffinity = &core.PodAntiAffinity{
-			RequiredDuringSchedulingIgnoredDuringExecution: []core.PodAffinityTerm{antiAffinityTerm},
-		}
-	}
-	return nil
-}
-
 const applyRawSpecTimeoutSeconds = 20
 
 func (k *kubernetesClient) applyRawK8sSpec(
@@ -1263,7 +1024,9 @@ func (k *kubernetesClient) ensureService(
 			return errors.Annotatef(err, "configuring devices for %s", appName)
 		}
 	}
-	if err := processConstraints(&workloadSpec.Pod.PodSpec, appName, params.Constraints); err != nil {
+	if err := k8sapplication.ApplyConstraints(
+		&workloadSpec.Pod.PodSpec, appName, params.Constraints, k8sapplication.ConfigureConstraint,
+	); err != nil {
 		return errors.Trace(err)
 	}
 
@@ -1460,18 +1223,6 @@ func (k *kubernetesClient) configureDevices(unitSpec *workloadSpec, devices []de
 		} else {
 			unitSpec.Pod.NodeSelector = nodeSelector
 		}
-	}
-	return nil
-}
-
-func configureConstraint(pod *core.PodSpec, constraint, value string) error {
-	for i := range pod.Containers {
-		resources := pod.Containers[i].Resources
-		err := mergeConstraint(constraint, value, &resources)
-		if err != nil {
-			return errors.Annotatef(err, "merging constraint %q to %#v", constraint, resources)
-		}
-		pod.Containers[i].Resources = resources
 	}
 	return nil
 }
@@ -2827,22 +2578,6 @@ func mergeDeviceConstraints(device devices.KubernetesDeviceParams, resources *co
 	// - https://kubernetes.io/docs/tasks/manage-gpus/scheduling-gpus/#clusters-containing-different-types-of-nvidia-gpus
 	resources.Limits[resourceName] = *resource.NewQuantity(device.Count, resource.DecimalSI)
 	resources.Requests[resourceName] = *resource.NewQuantity(device.Count, resource.DecimalSI)
-	return nil
-}
-
-func mergeConstraint(constraint string, value string, resources *core.ResourceRequirements) error {
-	if resources.Limits == nil {
-		resources.Limits = core.ResourceList{}
-	}
-	resourceName := core.ResourceName(constraint)
-	if v, ok := resources.Limits[resourceName]; ok {
-		return errors.NotValidf("resource limit for %q has already been set to %v!", resourceName, v)
-	}
-	parsedValue, err := resource.ParseQuantity(value)
-	if err != nil {
-		return errors.Annotatef(err, "invalid constraint value %q for %v", value, constraint)
-	}
-	resources.Limits[resourceName] = parsedValue
 	return nil
 }
